Upated tests
