+++ src/packages/piral-tracking/src/types.ts
@@ -61,3 +61,30 @@ export interface PiralTrackEndFrameEvent {
   properties: any;
   measurements: any;
 }
+
+/**
+ * Defines the provided set of tracking and telemetry Pilet API extensions.
+ */
+export interface PiralTrackingApi {
+  /**
+   * Tracks a simple (singular) event at the current point in time.
+   * @param name The name of the event to track.
+   * @param properties The optional tracking properties to submit.
+   * @param measurements The optional tracking measurements to submit.
+   */
+  trackEvent(name: string, properties?: any, measurements?: any): void;
+  /**
+   * Tracks an exception event at the current point in time.
+   * @param exception The Error from a catch clause, or the string error message.
+   * @param properties The optional tracking properties to submit.
+   * @param measurements The optional tracking measurements to submit.
+   * @param severityLevel The optional severity level of error.
+   */
+  trackError(error: Error | string, properties?: any, measurements?: any, severityLevel?: SeverityLevel): void;
+  /**
+   * Starts tracking an event frame at the current point in time.
+   * @param name The name of the event to start tracking.
+   * @returns The method to use for ending the current event frame.
+   */
+  trackFrame(name: string): Tracker;
+}
